Re-export dependent crates which the `probers` macro uses
When decorating a provider trait with `probers`, a lot of infrastructure
code gets generated to support probing.  Some of that code uses types
from the `failure` and `once_cell` crates.  By re-exporting those
crates, the macro is able to generate code that uses those re-exports,
without requiring the users of this macro to add those dependencies
themselves.