Create new `codegen` crate and start to refactor macros into it
I'm exploring another approach, in which I use some build-time code
generation to generate a very small thunking binary which the Rust code
can link to.  It's a much more complex approach and may end up being too
heavy, but if it works it'll mean the probers are statically compiled
into the Rust binaries which would be vastly preferable.