Substantially refactor macros moving most of the impl to codegen
The `probers-macros` crate is now providing a very thin wrapper around
the actual implementation in `probers-codegen`.  This means the
build-time code gen of the C wrappers and the compile time code-gen (via
proc macros) of the actual Rust probing calls are unified in
`probers-codegen`, where they share some of the same code.