Factor trait code generation out from parsing
This is the new architecture going forward.  The code that looks at the
Rust AST and constructs the provider and probe specifications is in one
place, and shared by all the implementations.  Separate code generate
modules for each implementation each consume those specs and produce
their own Rust code dependng on the implementation.

This commit makes this refactor for the most complicated case,
applying the `probers` attribute on a provider trait.  Next up are the
other two macros, then build-time.