Refactor STap/USDT code generation even more.
No new functionality, but I'm trying to find a structure for the
existing generator that will fit well with my ideas for the
implementation that relies on build-time generation of native libraries.
I like this current structure, the only thing I don't like is the code
generation using the `quote` crate is not as easy to work with visually.