Refactor APIs to create specs from tokens and generate from specs
Before the line was a bit blurry between the `TokenStream` the compiler
provides to the proc macros, the parsing of those streams into the
expected `syn` types, the processing of those `syn` types into things
like `ProbeSpecification` and `ProbeSpecification`, and then the
creation of generator types to generate the resulting code.

Now I'm making that separation much more stark.

Proc macros always take a `TokenStream`, which is consumed by the
factory method for the corresponding spec type.

Generators always consume the corresponding spec type, and generate an
output `TokenStream`.

I haven't applied this model to all the macros yet but I did the major
one, provider traits.  Not only does this clarify the responsibilities
between the various components, but it also allows each component to
take ownership of its input rather than holding a reference and cloning
the bits it needs.