Rename the `stap_usdt` generators to `dynamic`
At code generation time, there's nothing specific to `libstapsdt` in the
`stap_usdt` generator.  It will work for any implementation of the
dynamic tracing api in `probers-core`.  So to clear things up I've
renamed the module accordingly.

Similarly, the work I'm doing to support compile-time SystemTap support
will be mostly in the `static` module.  The generated C bindings are the
only things that are specific to any one tracing implementation.