Substantially refactor the directory structure into a virtual workspace
I was having a lot of problems with the top-level `probers` crate being
in the root of the workspace.  I looked at other similar crates and they
use a virtual workspace which has no package of its own.  I've done that
too.  The main `probers` crate used to be in `./`, but now it's in
`./probers`, a peer of all the other crates.  Dependency resolution
still works as before, and I don't have to manually include or exclude
paths in the workspace