Implement a sophisitcated build-time tracing implementation selection
Now all of the dependencies will compile on all platforms, even if
they're not supported.  For example, `libstapstd-sys` should build on
Mac and Windows, even though the code isn't supported there.  Whether
the crate compiles with a working implementation or just an empty shell
is indicated in a variable which is passed to dependent crates.

This allows a much more sophisticated build-time resolution of
dependencies in `build.rs` on `probers`.  By default it will use the
`stap` implementation if it compiled successfully, but will fall back to
using the `noop` implementation if anything prevented `stap` from
working.

This is just the start.  Once I implement a static form of `stap`
tracing that will become the default.  On windows, Mac, and BSD
presumably there will be other default implementations