Factor out `noop` into `dyn-noop` and implement generator switching
Since the `build.rs` in `probers` decides which implementation to use,
and includes or excludes runtime components based on that decision, when
the proc macros start to generate probing code they need to know what
decision was made.  That's now been implemented by storing that
information in a JSON file in the `$OUT_DIR` of the crate being compiled
with probers support (in this case, `probers` itself but the concept
applies to any downstream crate).