Remove old error type and use consistent new one throughout
The code was a mix of `Fallible` and `ProbeResult`, with no real pattern
for which was which.  Now all errors are reported via a `ProbersResult`,
which exposes a `ProbersError` type which also implements `Fail` from
the `failure` crate so it's easier to work with a mix of different error
types.

It also implements `Send + Sync` and can contain `Span`s to preserve
context in macro errors for the compiler to show to the user.