Add an example crate completely separate from `probers`.
The `probers` crate has several example files in its `examples/` folder,
all of which are compiled and run when `cargo test` is performed.  But I
needed examples that are in an entirely separate crate, to exercise the
macros in the way they'll be used by actual users.

Turns out this was a good call, because my simple "Hello World" sample
failed with a meaningless compiler error.  Oops.