Refactor generator creation and add a `Disabled` option
I was using `NativeNoOp` for the disabled case, but the problem is that
even that code generator uses some of the runtime types in `tracers`.
To me, "disabled" should mean that the caller crate doesn't even need a
run-time dependency on `tracers`; only the `macros` crate` but that's
only a compile-time dependency.