Separate out dynamic APIs and make them conditional
This change adds a lot more conditonal compilation.  If dynamic tracing
isn't enabled, none of the dynamic APIs will even be compiled in, so if
due to a bug one of the dynamic generators is used the result will be an
immediately obvious compile error.

I've also made any dependencies that are specific to dynamic support
entirely conditional on that being enabled.