Rename the `native` tracing implementations to `static`
I initially used `native` because `static` is a reserved word in Rust so
it can't be used as a module name.  But I couldn't get over the fact
that `static` is the right word for this, and it turned out there aren't
that many places where I had to escape `static` as `r#static`, and the
resulting clarify in the API and docs is well worth it.