Refactor code gen into two basic generators: static and dynamic
I realized there was a bit of a diffusion of responsibility in the code
generators, but in terms of actually logic there are two different
kinds: static and dynamic.  Even `disabled` is just a special case of
`static`.  So I removed the special `noop` and `disabled` generators
entirely and replaced them with a common `static` generator
parameterized with the implementation details.