Add additional enums describing tracing implementation details
I found myself writing a lot of conditional code based on bools, because
I didn't have a way to write code that applied to, for example, "all
dynamic impls" or "all stap impls".  I added additional enums for the
tracing targets and the tracing types, to help with this.  It's made
quite a bit of code more clear, and also using Rust's compiler to ensure
the match expressions are exhaustive so when I add another impl down the
line I'll get compile errors every place I need to update