Implement generation of Rust bindings for native wrappers
My initial implementation generated the Rust bindings in `build.rs` and
saved them as a file, which the proc macros would then `include!` into
the implementation module.  But that felt ugly and awkward and it wasn't
at all clear why some but not all Rust code should be generated with a
Askama template and the rest with the proc macros.

So, I'm removing the generation of Rust bindings at build time.  Of
course the native bindings are still generated this way, but the proc
macros will know how to generate Rust bindings themselves.  This fits
much more cleanly in the design, and it removes a bit of the "spooky
action at a distance" feel from the previous implementation.