Generate the same basic probing code for `NoOp` and `Stap`
Now the no-op impl and the stap impl (and, in the future, the other
static impls) use the same basic approach, so the code generation code
is almost all shared, and the `probe!` macro doesn't need to know
whether it's generating for `noop` or a real impl because the code it
should produce is the same.

There are still compile errors in the generated code so this change
isn't being pushed.