Fix provider spec hash computation to avoid variations due to tokens
When tokens come into a proc macro, sometimes there are subtle
variations like an extra whitespace in parens, such that the hashes of
the string representation are different.

Fortunately it turns out `ItemTrait` implements `Hash`, and since
everything we know about a provider comes from the `ItemTrait`, it's
safe to use that as the hash.