Introduce explicit `wrap` function for static probe types
I had hoped this wouldn't be necessary, but because the `probe!` macro
doesn't know anything about the probe's argument types, it can't always
determine the correct `ProbeArgType` implementation to use to wrap the
arguments, which leads to compiler errors.

So instead I've added a `__wrap` method for each probe method, which
takes as input all of the probe's parameters, and returns a tuple of
`ProbeArgWrapper` types matching the probe's parameters.  This is
actually rather similar to the code used to support dynamic probes.

My hope is that this will mostly compile out, particularly for the
primitive types which don't actually need any kind of wrapping.