Remove explicit static linking to native libraries from build.rs
There's no reason to force the linking with every wrapper library at
build time.  As a result of doing this, all probe provider names had to
be globally unique; that is, a test or an example couldn't use the same
provider and probe name as anything else in the dependency chain.

That's not necessary, and in fact it's less efficient.  Since the
`tracer` attribute automatically adds a `link` attribute to force
linking with the appropriate library, it wasn't necessary to instruct
`rustc` to link with the generated libs at build time.