Extend probe benchmarks to also test probes firing (stap only)
This is a very interesting result.  The overhead for a probe call when
using stap and the probe is enabled is ~500ps, which is the time spent
checking the semaphore's value.  With the probes enabled with
`funccount` which does nothing but tally up the number of times each
probe is called, the no-arg and int-arg probes become 1000x slower, on
the order of 500ns.  The string-arg probes have more overhead because of
the conversion to C-style strings; the slowest one is over 1us with 12
string args.

I must admit I'm disappointed by the significance of the performance
impact.